fix: fix the rest of the tests
