+++ apps/barry/tests/modules/marketplace/profiles/database/ProfileRepository.test.ts
@@ -1,104 +1,6 @@
-import type { ProfilesSettings } from "@prisma/client";
-
-import {
-    ProfileMessageRepository,
-    ProfileRepository,
-    ProfilesSettingsRepository
-} from "../../../../src/modules/marketplace/dependencies/profiles/database/index.js";
-
-import { mockChannel } from "@barry/testing";
-import { mockProfile } from "./mocks/profile.js";
-import { prisma } from "../../../mocks/index.js";
-
-describe("ProfilesSettingsRepository", () => {
-    const guildID = "68239102456844360";
-
-    let repository: ProfilesSettingsRepository;
-    let settings: ProfilesSettings;
-
-    beforeEach(() => {
-        repository = new ProfilesSettingsRepository(prisma);
-        settings = {
-            channelID: mockChannel.id,
-            enabled: true,
-            guildID: guildID,
-            lastMessageID: null
-        };
-    });
-
-    describe("create", () => {
-        it("should create a new profiles settings record for the specified guild", async () => {
-            await repository.create(guildID);
-
-            expect(prisma.profilesSettings.create).toHaveBeenCalledOnce();
-            expect(prisma.profilesSettings.create).toHaveBeenCalledWith({ data: { guildID } });
-        });
-    });
-
-    describe("get", () => {
-        it("should retrieve the profiles settings record for the specified guild", async () => {
-            vi.mocked(prisma.profilesSettings.findUnique).mockResolvedValue(settings);
-
-            const entity = await repository.get(guildID);
-
-            expect(entity).toEqual(settings);
-            expect(prisma.profilesSettings.findUnique).toHaveBeenCalledOnce();
-            expect(prisma.profilesSettings.findUnique).toHaveBeenCalledWith({ where: { guildID } });
-        });
-
-
-        it("should return null when no profiles settings record is found", async () => {
-            vi.mocked(prisma.profilesSettings.findUnique).mockResolvedValue(null);
-
-            const entity = await repository.get(guildID);
-
-            expect(entity).toBeNull();
-        });
-    });
-
-    describe("getOrCreate", () => {
-        it("should retrieve the profiles settings record if found", async () => {
-            vi.mocked(prisma.profilesSettings.findUnique).mockResolvedValue(settings);
-
-            const entity = await repository.getOrCreate(guildID);
-
-            expect(entity).toEqual(settings);
-            expect(prisma.profilesSettings.findUnique).toHaveBeenCalledOnce();
-            expect(prisma.profilesSettings.create).not.toHaveBeenCalled();
-        });
-
-        it("should create a new profiles settings record if not found", async () => {
-            vi.mocked(prisma.profilesSettings.findUnique).mockResolvedValue(null);
-            vi.mocked(prisma.profilesSettings.create).mockResolvedValue(settings);
-
-            const entity = await repository.getOrCreate(guildID);
-
-            expect(entity).toEqual(settings);
-            expect(prisma.profilesSettings.findUnique).toHaveBeenCalledOnce();
-            expect(prisma.profilesSettings.create).toHaveBeenCalledOnce();
-        });
-    });
-
-    describe("upsert", () => {
-        it("should upsert the provided profiles settings record", async () => {
-            await repository.upsert(guildID, {
-                lastMessageID: "91256340920236565"
-            });
-
-            expect(prisma.profilesSettings.upsert).toHaveBeenCalledOnce();
-            expect(prisma.profilesSettings.upsert).toHaveBeenCalledWith({
-                create: {
-                    guildID: guildID,
-                    lastMessageID: "91256340920236565"
-                },
-                update: {
-                    lastMessageID: "91256340920236565"
-                },
-                where: { guildID }
-            });
-        });
-    });
-});
+import { ProfileRepository } from "../../../../../src/modules/marketplace/dependencies/profiles/database/index.js";
+import { mockProfile } from "../mocks/profile.js";
+import { prisma } from "../../../../mocks/index.js";
 
 describe("ProfileRepository", () => {
     const guildID = "68239102456844360";
@@ -257,60 +159,6 @@ describe("ProfileRepository", () => {
                 },
                 where: { userID }
             });
-        });
-    });
-});
-
-describe("ProfileMessageRepository", () => {
-    const guildID = "68239102456844360";
-    const messageID = "91256340920236565";
-    const userID = "257522665441460225";
-
-    let repository: ProfileMessageRepository;
-
-    beforeEach(() => {
-        repository = new ProfileMessageRepository(prisma);
-    });
-
-    describe("create", () => {
-        it("should create a new profile message record", async () => {
-            await repository.create(messageID, guildID, userID);
-
-            expect(prisma.profileMessage.create).toHaveBeenCalledOnce();
-            expect(prisma.profileMessage.create).toHaveBeenCalledWith({
-                data: {
-                    guildID,
-                    messageID,
-                    userID
-                }
-            });
-        });
-    });
-
-    describe("getLatest", () => {
-        it("should return the latest profile message for the specified member", async () => {
-            vi.mocked(prisma.profileMessage.findFirst).mockResolvedValue({
-                guildID,
-                messageID,
-                userID
-            });
-
-            const entity = await repository.getLatest(guildID, userID);
-
-            expect(entity).toEqual({ guildID, messageID, userID });
-            expect(prisma.profileMessage.findFirst).toHaveBeenCalledOnce();
-            expect(prisma.profileMessage.findFirst).toHaveBeenCalledWith({
-                orderBy: { messageID: "desc" },
-                where: { guildID, userID }
-            });
-        });
-
-        it("should return null if the user has not yet posted a message in the specified guild", async () => {
-            vi.mocked(prisma.profileMessage.findFirst).mockResolvedValue(null);
-
-            const entity = await repository.getLatest(guildID, userID);
-
-            expect(entity).toBeNull();
         });
     });
 });
