+++ apps/barry/tests/modules/marketplace/requests/database/RequestRepository.test.ts
@@ -1,13 +1,7 @@
-import { type RequestsSettings, Prisma, RequestStatus } from "@prisma/client";
-import {
-    RequestMessageRepository,
-    RequestRepository,
-    RequestsSettingsRepository
-} from "../../../../src/modules/marketplace/dependencies/requests/database/index.js";
-
-import { mockChannel } from "@barry/testing";
-import { mockRequest } from "./mocks/request.js";
-import { prisma } from "../../../mocks/index.js";
+import { Prisma, RequestStatus } from "@prisma/client";
+import { RequestRepository } from "../../../../../src/modules/marketplace/dependencies/requests/database/index.js";
+import { mockRequest } from "../mocks/request.js";
+import { prisma } from "../../../../mocks/index.js";
 
 describe("RequestRepository", () => {
     const guildID = "68239102456844360";
@@ -271,114 +265,6 @@ describe("RequestRepository", () => {
             await expect(() => repository.upsert(userID, data)).rejects.toThrowError(
                 "Missing required fields."
             );
-        });
-    });
-});
-
-describe("RequestMessageRepository", () => {
-    const guildID = "68239102456844360";
-    const messageID = "91256340920236565";
-    const requestID = 1;
-
-    let repository: RequestMessageRepository;
-
-    beforeEach(() => {
-        repository = new RequestMessageRepository(prisma);
-    });
-
-    describe("create", () => {
-        it("should create a new request message record", async () => {
-            await repository.create(messageID, guildID, requestID);
-
-            expect(prisma.requestMessage.create).toHaveBeenCalledOnce();
-            expect(prisma.requestMessage.create).toHaveBeenCalledWith({
-                data: {
-                    guildID,
-                    messageID,
-                    requestID
-                }
-            });
-        });
-    });
-
-    describe("getLatest", () => {
-        it("should return the latest request message for the specified member", async () => {
-            vi.mocked(prisma.requestMessage.findFirst).mockResolvedValue({
-                guildID,
-                messageID,
-                requestID
-            });
-
-            const entity = await repository.getLatest(guildID, requestID);
-
-            expect(entity).toEqual({ guildID, messageID, requestID });
-            expect(prisma.requestMessage.findFirst).toHaveBeenCalledOnce();
-            expect(prisma.requestMessage.findFirst).toHaveBeenCalledWith({
-                orderBy: { messageID: "desc" },
-                where: { guildID, requestID }
-            });
-        });
-
-        it("should return null if the user has not yet posted a message in the specified guild", async () => {
-            vi.mocked(prisma.requestMessage.findFirst).mockResolvedValue(null);
-
-            const entity = await repository.getLatest(guildID, requestID);
-
-            expect(entity).toBeNull();
-        });
-    });
-});
-
-describe("RequestsSettingsRepository", () => {
-    const guildID = "68239102456844360";
-
-    let repository: RequestsSettingsRepository;
-    let settings: RequestsSettings;
-
-    beforeEach(() => {
-        repository = new RequestsSettingsRepository(prisma);
-        settings = {
-            channelID: mockChannel.id,
-            enabled: true,
-            guildID: guildID,
-            lastMessageID: null,
-            minCompensation: 50
-        };
-    });
-
-    describe("getOrCreate", () => {
-        it("should return the requests settings for a guild", async () => {
-            vi.mocked(prisma.requestsSettings.upsert).mockResolvedValue(settings);
-
-            const entity = await repository.getOrCreate(guildID);
-
-            expect(entity).toEqual(settings);
-            expect(prisma.requestsSettings.upsert).toHaveBeenCalledOnce();
-            expect(prisma.requestsSettings.upsert).toHaveBeenCalledWith({
-                create: { guildID },
-                update: {},
-                where: { guildID }
-            });
-        });
-    });
-
-    describe("upsert", () => {
-        it("should upsert the provided level up settings record", async () => {
-            await repository.upsert(guildID, {
-                lastMessageID: "91256340920236565"
-            });
-
-            expect(prisma.requestsSettings.upsert).toHaveBeenCalledOnce();
-            expect(prisma.requestsSettings.upsert).toHaveBeenCalledWith({
-                create: {
-                    guildID: guildID,
-                    lastMessageID: "91256340920236565"
-                },
-                update: {
-                    lastMessageID: "91256340920236565"
-                },
-                where: { guildID }
-            });
         });
     });
 });
