+++ apps/barry/tests/modules/marketplace/requests/events/contactRequest.test.ts
@@ -45,7 +45,7 @@ describe("Contact (InteractionCreate) Event", () => {
 
             delete data.guild_id;
             const interaction = new MessageComponentInteraction(data, event.client, vi.fn());
-            const settingsSpy = vi.spyOn(event.module.requestsSettings, "getOrCreate");
+            const settingsSpy = vi.spyOn(event.module.settings, "getOrCreate");
 
             await event.execute(interaction);
 
@@ -54,7 +54,7 @@ describe("Contact (InteractionCreate) Event", () => {
 
         it("should ignore if the interaction is not of type 'MessageComponent'", async () => {
             const interaction = new PingInteraction(mockPingInteraction, event.client, vi.fn());
-            const settingsSpy = vi.spyOn(event.module.requestsSettings, "getOrCreate");
+            const settingsSpy = vi.spyOn(event.module.settings, "getOrCreate");
 
             await event.execute(interaction);
 
@@ -69,7 +69,7 @@ describe("Contact (InteractionCreate) Event", () => {
             });
 
             const interaction = new MessageComponentInteraction(data, event.client, vi.fn());
-            const settingsSpy = vi.spyOn(event.module.requestsSettings, "getOrCreate");
+            const settingsSpy = vi.spyOn(event.module.settings, "getOrCreate");
 
             await event.execute(interaction);
 
@@ -83,7 +83,7 @@ describe("Contact (InteractionCreate) Event", () => {
             });
 
             const interaction = new MessageComponentInteraction(data, event.client, vi.fn());
-            const settingsSpy = vi.spyOn(event.module.requestsSettings, "getOrCreate");
+            const settingsSpy = vi.spyOn(event.module.settings, "getOrCreate");
 
             await event.execute(interaction);
